Move L10n's get method to a separate file
